rename: style 관련 폴더 구조 변경 및 theme.ts 내부 요소 분리 (#2) - theme.ts에 있던 color와 font를 다른 파일로 분리 - global style 관련 파일을 `global` 폴더로 이동 - theme 관련 파일을 'theme' 폴더로 이동
