+++ src/styles/index.ts
@@ -1,0 +1,4 @@
+export * from '@/styles/global/GlobalFont';
+export * from '@/styles/global/GlobalStyle';
+
+export * from '@/styles/theme/theme';
