+++ src/styles/styled.d.ts
@@ -1,9 +1,9 @@
+import { ColorTypes, FontTypes } from '@/styles/theme/theme';
 import 'styled-components';
-import { ColorsTypes, FontsTypes } from './theme';
 
 declare module 'styled-components' {
   export interface DefaultTheme {
-    colors: ColorsTypes;
-    fonts: FontsTypes;
+    color: ColorTypes;
+    font: FontTypes;
   }
 }
